refactor timeWindow code

+++ header.go
@@ -25,7 +25,7 @@ type header struct {
 func (h header) MarshalBinary() ([]byte, error) {
 	buf := make([]byte, headerSize)
 	copy(buf[:7], h.signature[:])
-	buf[7] = h.encryption
+	buf[7] = uint8(h.encryption)
 	binary.LittleEndian.PutUint32(buf[8:12], h.version)
 	binary.LittleEndian.PutUint64(buf[12:20], h.seq)
 	binary.LittleEndian.PutUint64(buf[20:28], uint64(h.count))
@@ -39,7 +39,7 @@ func (h header) MarshalBinary() ([]byte, error) {
 // UnmarshalBinary de-serializes header from binary data
 func (h *header) UnmarshalBinary(data []byte) error {
 	copy(h.signature[:], data[:7])
-	h.encryption = data[7]
+	h.encryption = int8(data[7])
 	h.seq = binary.LittleEndian.Uint64(data[12:20])
 	h.count = int64(binary.LittleEndian.Uint64(data[20:28]))
 	h.windowIdx = int32(binary.LittleEndian.Uint32(data[28:32]))
